Fix for it triggering antivirus
This fixed it triggering anti virus on my end.

mounted() { //this will run once the app.js is mounted fully ready too go
    window.addEventListener('message', this.onMessage) //creates an event lister so it will listen too any nui events that run
  },

That was line 14 - 16 before

mounted() { //this will run once the app.js is mounted fully ready too go
    window.addEventListener('message', this.onMessage); //creates an event lister so it will listen too any nui events that run
  },

line 14-16 now. Looked at bcc-scenes and the only diff was the ; at the end of the event listener added that and it isnt triggering anti virus anymore. Do not know why that fixed it but it did